+++ src/main/java/com/github/mozvip/subtitles/MovieSubtitlesFinder.java
@@ -1,10 +1,11 @@
 package com.github.mozvip.subtitles;
 
+import java.math.BigDecimal;
 import java.util.Locale;
 
 public interface MovieSubtitlesFinder {
 	
-	public RemoteSubTitles downloadMovieSubtitles(String movieName, int year, String release, float fps, Locale locale) throws Exception;
+	public RemoteSubTitles downloadMovieSubtitles(String movieName, int year, String release, BigDecimal fps, Locale locale) throws Exception;
 	
 
 }
